[FIX] Edge cases that crashed the ui solved
There where 2 edge cases that crashed the GUI within the js, one is when you tried to change the possition of a another line while you are adding a line with this module feature of adding them in the middle before setting its product, this hardcrashed the gui and you had to reload, i solved it by disabling handlers when a new line added with this module until it is either deleted or a product assigned to it. As for the other crash it was when you added a new line with the vanilla odoo feature, that if you tried to move other line it soft crashed the gui, you did not had to reload but it still messed up the styles and a error warning jumped, this has been completly solved without tricks.